# user.go
package main

import (
	"fmt"
	"net/http"
	"strings"

	"github.com/gin-gonic/gin"
	"github.com/rs/zerolog/log"
)

// Database
var Users []ModelUser
var UserAddress = make(map[string][]RequestUserAddressToAdd)

// User Table
type ModelUser struct {
	Name     string `json:"name" validate:"required"`
	Username string `json:"username" validate:"required"`
	Password string `json:"password" validate:"required"`
}

// DTO - Data Transfer Object
// Login Request
type RequestLogin struct {
	Username string `json:"username" validate:"required"`
	Password string `json:"password" validate:"required"`
}

// Request to Add Address
type RequestUserAddressToAdd struct {
	Add string `json:"address"`
}

// Request to Get User Profile
type ResponseUserProfile struct {
	Name     string                    `json:"name" validate:"required"`
	Username string                    `json:"username" validate:"required"`
	Address  []RequestUserAddressToAdd `json:"address,omitempty"`
	Books    []ModelBook               `json:"books,omitempty"`
	Cart     []CheckOut                `json:"cart,omitempty"`
	Orders   []Orders                  `json:"orders,omitempty"`
}

// Create User
// @Schemes http
// @Description Create User
// @Tags User
// @Accept json
// @Produce json
// @Param			request_body	body		ModelUser	true	"Request Body"
// @Success 200 {object} SuccessDTO
// @Failure 409 {object} ErrorDTO
// @Router /signup [post]
func CreateUser(c *gin.Context) {
	//Request User Data
	var user ModelUser
	//BindJSON
	jsonRes := Bindjson(c, &user)
	if jsonRes {
		return
	}
	//Validate JSON
	jsonValid := ValidateJson(c, &user)
	if jsonValid {
		return
	}
	//Print Incoming Request
	log.Info().Msgf("Request URL :: %s --- Request Method :: %s  --- Request Body :: %+v", c.Request.URL, c.Request.Method, user)
	//Check Already Exists Book Record
	if isUserExists(user.Username) {
		c.JSON(http.StatusConflict, ErrorDTO{
			ErrorCode:    fmt.Sprintf("%d", http.StatusConflict),
			ErrorMessage: fmt.Sprintf("Username - %s already exists", user.Username),
		})
		log.Warn().Msgf("Username - %s already exists", user.Username)
		return
	}
	// Save user in table
	Users = append(Users, user)
	//Response
	c.JSON(http.StatusOK, SuccessDTO{
		SuccessCode:    fmt.Sprintf("%d", http.StatusOK),
		SuccessMessage: "New User Created Successfully",
		CustomMessage: map[string]string{
			"username": user.Username,
			"password": strings.Repeat("*", len(user.Password)),
		},
	})
}

// Login User
// @Schemes http
// @Description Login User
// @Tags User
// @Accept json
// @Produce json
// @Param			request_body	body		RequestLogin	true	"Request Body"
// @Success 200 {object} SuccessDTO
// @Failure 401 {object} ErrorDTO
// @Failure 404 {object} ErrorDTO
// @Router /signin [post]
func LoginUser(c *gin.Context) {
	//Request for User Login
	var user RequestLogin
	//BindJSON
	jsonRes := Bindjson(c, &user)
	if jsonRes {
		return
	}
	//Validate JSON
	jsonValid := ValidateJson(c, &user)
	if jsonValid {
		return
	}
	//Print Incoming Request
	log.Info().Msgf("Request URL :: %s --- Request Method :: %s  --- Request Body :: %+v", c.Request.URL, c.Request.Method, user)
	//Check Already Exists User Record
	if !isUserExists(user.Username) {
		//Response
		c.JSON(http.StatusNotFound, ErrorDTO{
			ErrorCode:    fmt.Sprintf("%d", http.StatusNotFound),
			ErrorMessage: fmt.Sprintf("Username - %s don't have a account", user.Username),
		})
		//Print Incoming Request
		log.Warn().Msgf("Username - %s don't have a account", user.Username)
		return
	}
	//Check user credentials
	for i := 0; i < len(Users); i++ {
		if Users[i].Username == user.Username && Users[i].Password == user.Password {
			//Generate Token
			token, err := GenerateToken(user.Username)
			if err != nil {
				c.JSON(http.StatusInternalServerError, ErrorDTO{
					ErrorCode:    fmt.Sprintf("%d", http.StatusInternalServerError),
					ErrorMessage: "Failed to generate token",
				})
				return
			}
			log.Info().Msgf("Login SuccessFully for Username - %s and Password - %s", user.Username, strings.Repeat("*", len(user.Password)))
			c.JSON(http.StatusOK, SuccessDTO{
				SuccessCode:    fmt.Sprintf("%d", http.StatusOK),
				SuccessMessage: fmt.Sprintf("Login SuccessFully for Username - %s", user.Username),
				CustomMessage:  token,
			})
			return
		}
	}
	c.JSON(http.StatusUnauthorized, ErrorDTO{
		ErrorCode:    fmt.Sprintf("%d", http.StatusUnauthorized),
		ErrorMessage: "Invalid username or password",
	})
}

// Add User Address
// @Schemes
// @Description Add User Address
// @Tags User
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param	request_body	body RequestUserAddressToAdd	true	"Request Body"
// @Success 200 {object} SuccessDTO
// @Failure 409 {object} ErrorDTO
// @Router /user/address [post]
func UserAdd(c *gin.Context) {
	//To Store Active Username
	activeUsername := c.GetString("username")
	//Declare DTO for Book
	var add RequestUserAddressToAdd
	//BindJSON
	jsonRes := Bindjson(c, &add)
	if jsonRes {
		return
	}
	//Validate JSON
	jsonValid := ValidateJson(c, &add)
	if jsonValid {
		return
	}
	//Print Incoming Request
	log.Info().Msgf("Username - %s :: Requested :: Request URL :: %s --- Request Method :: %s  --- Request Body :: %+v", activeUsername, c.Request.URL, c.Request.Method, add)
	// Address
	if isAddExists(add, activeUsername) {
		//Response
		c.JSON(http.StatusConflict, ErrorDTO{
			ErrorCode:    fmt.Sprintf("%d", http.StatusConflict),
			ErrorMessage: fmt.Sprintf("Address - %s already exists", add.Add),
		})
		//Print Incoming Request
		log.Warn().Msgf("Address - %s already exists", add.Add)
		return
	}
	if val, ok := UserAddress[activeUsername]; ok {
		UserAddress[activeUsername] = append(val, add)
	} else {
		UserAddress[activeUsername] = []RequestUserAddressToAdd{
			{Add: add.Add},
		}
	}
	//Response
	c.JSON(http.StatusOK, SuccessDTO{
		SuccessCode:    fmt.Sprintf("%d", http.StatusOK),
		SuccessMessage: "User Address added Successfully",
		CustomMessage:  UserAddress[activeUsername],
	})
}

// Get User Profile By Username
// @Schemes
// @Description Get User Profile By Username
// @Tags User
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param username	path	string	true "username"
// @Success 200 {object} SuccessDTO "Success"
// @Failure 404 {object} ErrorDTO "Error"
// @Router /user/profile/{username} [get]
func GetProfile(c *gin.Context) {
	//Username
	activeUsername := c.Params.ByName("username")
	//Print Incoming Request
	log.Info().Msgf("Request URL :: %s --- Request Method :: %s --- Username  :: %v", c.Request.URL, c.Request.Method, activeUsername)
	//IsUserExists
	if !isUserExists(activeUsername) {
		//Response
		c.JSON(http.StatusNotFound, ErrorDTO{
			ErrorCode:    fmt.Sprintf("%d", http.StatusNotFound),
			ErrorMessage: fmt.Sprintf("Username - %s does not have an account", activeUsername),
		})
		//Print Incoming Request
		log.Warn().Msgf("Username - %s does not have an account", activeUsername)
		return
	}
	//Response
	c.JSON(http.StatusOK, SuccessDTO{
		SuccessCode:    fmt.Sprintf("%d", http.StatusOK),
		SuccessMessage: fmt.Sprintf("Profile of Username %s", activeUsername),
		CustomMessage: ResponseUserProfile{
			Name:     GetName(activeUsername),
			Username: activeUsername,
			Address:  UserAddress[activeUsername],
			Books:    BooksData[activeUsername],
			Cart:     CartData[activeUsername],
			Orders:   OrderData[activeUsername],
		},
	})
}

// Check Is User Exists
func isUserExists(username string) bool {
	for i := 0; i < len(Users); i++ {
		if Users[i].Username == username {
			return true
		}
	}
	return false
}

// Check Is User Exists
func isAddExists(add RequestUserAddressToAdd, username string) bool {
	if val, ok := UserAddress[username]; ok {
		for i := 0; i < len(val); i++ {
			if val[i] == add {
				return true
			}
		}
	}
	return false
}

// Get Name
func GetName(username string) string {
	for i := 0; i < len(Users); i++ {
		if Users[i].Username == username {
			return Users[i].Name
		}
	}
	return ""
}